+++ src/OrleansRuntime/Scheduler/PoliciedScheduler/SchedulingStrategies/TestSchedulingStrategy.cs
@@ -96,7 +96,7 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
             return wig;
         }
 
-        public double FetchWorkItemMetric(WorkItemGroup workItem)
+        public long FetchWorkItemMetric(WorkItemGroup workItem)
         {
             throw new NotImplementedException();
         }
